+++ Assets/Scripts/UI/Component/AnimatedProcessbar.cs
@@ -4,12 +4,12 @@ using System.Collections;
 public class AnimatedProcessbar : MonoBehaviour, IProcessable {
 	[SerializeField] ProcessBar upBar;
 	[SerializeField] ProcessBar downBar;
-	public void SetProcess (double percent, bool force = false) {
+	public void SetProcess (float percent, bool force = false) {
 		upBar.SetProcess (percent, force);
 		downBar.SetProcess (percent, force);
 	}
 
-	public void InitProcess (double percent) {
+	public void InitProcess (float percent) {
 		upBar.InitProcess (percent);
 		downBar.InitProcess (percent);
 	}
